# pdlc/arsene-cms/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Arsène CMS API
  version: 1.0.0
  summary: >-
    The four Arsène operations that need server-side atomicity and therefore
    exist as real Supabase Edge Function endpoints, instead of a direct
    PostgREST read/write against a table. Everything else in the writer
    editor (draft field edits/autosave, taxonomy/categories, the asset
    library, alt-text edits, cover/body re-tagging after upload) is a direct
    PostgREST call against Postgres under RLS and is deliberately NOT part of
    this contract — see "What is out of scope" below.
  description: |
    ## What is in scope, and why only four endpoints

    Arsène's writer editor is a single-page app that talks to Postgres
    directly via PostgREST/RLS for nearly everything: autosaving a draft's
    title/body/meta fields once it exists, creating a league or category,
    browsing the shared asset library, and editing an image's `alt_text` or
    `role` after it has been uploaded. None of that needs a bespoke endpoint;
    RLS is the enforcement point and Postgres row shape is the contract.

    Four operations are different, because they have side effects that must
    happen exactly once, atomically, from the server, and cannot be expressed
    as a single-row PostgREST write:

    - `POST /v1/articles` — creates a new draft and takes its edit lock for
      the creating writer in the same call that **synchronously** records the
...
      metric with nothing to detect it.
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
  /v1/articles:
    post:
      operationId: createDraft
      tags: [Articles]
      summary: Create a new draft article
      description: |
        Creates a new `articles` row in `draft` status, takes its edit lock
        for the calling writer (AC-05 — the creator holds the lock from the
        instant the draft exists, with no separate `open` call needed right
        after create), and **synchronously** records one `draft_started`
        telemetry event (spec §4: `writer_id`, `article_id`, `started_at`) as
        part of the same request — never a fire-and-forget side call, so a
        draft can never exist without also being measured
        (`02-architecture.v1.md` §9).

        **Every field is optional.** All three body fields may be omitted,
        together or individually, to create a fully blank draft. If `title`
        is omitted (or empty), the server defaults it to "Sans titre" — a
        draft's title is never blank, so it always has something displayable
        in the writer's draft list before the writer has typed anything.
        `league_name`/`type_name` set the draft's initial taxonomy fields
        when provided; like every other draft field, both remain editable
        afterward via direct PostgREST `PATCH` writes under RLS (see "What
        is out of scope" above), not through this endpoint.

        **Not idempotent, and deliberately has no `Idempotency-Key`.** See
        "Idempotency" above for why a duplicate draft from a client retry is
        an acceptable, low-cost outcome here, unlike publish or upload.
      security:
        - supabaseWriterAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDraftRequest'
            examples:
              blankDraft:
                summary: writer clicks "New draft" with no starting values at all
                value: {}
              withStartingValues:
                summary: writer picks a league/type before typing a title
                value:
                  title: 'Pronos Ligue 1 - Journée 17'
                  league_name: Ligue 1
                  type_name: Pronos
      responses:
        '201':
          description: The draft was created and its edit lock is held by the calling writer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDraftResponse'
              examples:
                created:
                  summary: a new draft, locked by its creator
                  value:
                    article_id: a1a1a1a1-0000-4a2b-9c3d-000000000001
                    locked_by: e5e5e5e5-0000-4a2b-9c3d-eeeeeeeeeeee
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /v1/articles/{articleId}/open:
    post:
      operationId: openDraft
      tags: [Articles]
      summary: Reopen an existing draft, taking its edit lock
      description: |
        Attempts to take this draft's edit lock for the calling writer via
        the same staleness-checked compare-and-swap described in
        `adr/0003-draft-locking-mechanism.md` (free, already held by this
        writer, or stale-by-more-than-90s all succeed; actively held by
        someone else does not), and returns the draft's current saved
        content in the same call — the lock decision and the content read
        must be resolved together, from the server, so two writers opening
        the same draft at once cannot both believe they hold it.

        **No request body.** This operation takes no input beyond the path's
        `articleId`; send an empty JSON object (`{}`) or an empty body.

        **AC-05 — locked by someone else.** If the lock is currently held by
        another writer and is not yet stale, this returns `409 DRAFT_LOCKED`
        with `error.details.locked_by_writer_id` and
        `error.details.locked_by_display_name`, so the editor can name who
        to ask, exactly as `POST .../publish`'s own `DRAFT_LOCKED` response
        does.

        **Idempotent for the lock holder, not for anyone else.** See
        "Idempotency" above.
      security:
        - supabaseWriterAuth: []
      parameters:
        - $ref: '#/components/parameters/ArticleIdPathParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              description: Deliberately empty — this operation takes no input beyond `articleId`.
            examples:
              empty:
                summary: the only body this operation accepts
                value: {}
      responses:
        '200':
          description: The lock is now held by the calling writer; the draft's current content follows.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenDraftResponse'
              examples:
                reopened:
                  summary: an existing draft, now locked by the calling writer
                  value:
                    article_id: a1a1a1a1-0000-4a2b-9c3d-000000000001
                    title: 'Pronos Ligue 1 - Journée 17'
                    body_html: '<p>Notre analyse de la journée 17...</p>'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The draft is currently locked for editing by another writer (AC-05).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                draftLocked:
                  summary: another writer currently holds the edit lock and their lock has not gone stale
                  value:
                    error:
                      code: DRAFT_LOCKED
                      message: This draft is currently locked for editing by another writer.
                      details:
                        locked_by_writer_id: e5e5e5e5-0000-4a2b-9c3d-eeeeeeeeeeee
                        locked_by_display_name: Marie D.
                      request_id: 5e5e5e5e-5555-4a2b-9c3d-eeeeeeeeeeee
        default:
          $ref: '#/components/responses/UnexpectedError'
...
components:
  securitySchemes:
    supabaseWriterAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Supabase Auth session token for the calling writer's own account.
        Sent as `Authorization: Bearer <token>`. Required on every operation
        in this contract; there is no unauthenticated operation here (unlike
        the sibling `pronos` project's account-less player endpoints).
  parameters:
    ArticleIdPathParam:
      name: articleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: An `articles.id` value.
      example: a1a1a1a1-0000-4a2b-9c3d-000000000001
...
  responses:
    ValidationError:
      description: The request body or parameters failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_FAILED
              message: Request failed validation.
              details:
                fields:
                  - field: role
                    message: must be one of "cover", "body"
              request_id: 4b3c2d1e-0f1a-4b2c-8d3e-4f5a6b7c8d9e
    Unauthorized:
      description: Missing or invalid writer bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: A valid Supabase Auth bearer token is required for this endpoint.
              details: null
              request_id: 7d8e9f0a-1b2c-4d3e-9f4a-5b6c7d8e9f0a
    NotFound:
      description: The referenced article does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: No article was found matching the given id.
              details:
                article_id: 00000000-0000-0000-0000-000000000000
              request_id: 8e9f0a1b-2c3d-4e5f-8a9b-6c7d8e9f0a1b
    UnexpectedError:
      description: >-
        Any status code not explicitly documented above for this operation.
        Always parseable as the same `Error` envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred. Please retry.
              details: null
              request_id: 9f0a1b2c-3d4e-4f5a-8b9c-7d8e9f0a1b2c
  schemas:
...
    CreateDraftRequest:
      type: object
      description: >-
        All three fields are optional starting values for the new draft;
        omit any or all of them to get a blank draft. See the operation
        description for the "Sans titre" default and how these values are
        edited afterward.
      properties:
        title:
          type: string
          example: 'Pronos Ligue 1 - Journée 17'
        league_name:
          type: string
          description: >-
            The draft's initial league/competition taxonomy value (matched
            or created against `leagues.name` — see `01-decisions.md` #2 on
            near-duplicate names not being merged automatically).
          example: Ligue 1
        type_name:
          type: string
          description: The draft's initial content-type taxonomy value.
          example: Pronos
    CreateDraftResponse:
      type: object
      required: [article_id, locked_by]
      properties:
        article_id:
          type: string
          format: uuid
          description: >-
            The new `articles.id`, immediately usable for autosave PostgREST
            writes (title/body/taxonomy fields) and, eventually,
            `POST .../publish`.
          example: a1a1a1a1-0000-4a2b-9c3d-000000000001
        locked_by:
          type: string
          format: uuid
          description: >-
            The creating writer's own id. Always equal to the calling
            writer's id — the creator holds the edit lock from the instant
            the draft exists (AC-05), so a client never needs to call
            `POST .../open` immediately after `POST /v1/articles` for its
            own just-created draft.
          example: e5e5e5e5-0000-4a2b-9c3d-eeeeeeeeeeee
    OpenDraftResponse:
      type: object
      required: [article_id, title, body_html]
      properties:
        article_id:
          type: string
          format: uuid
          example: a1a1a1a1-0000-4a2b-9c3d-000000000001
        title:
          type: string
          description: The draft's current saved title (autosaved via PostgREST, not through this API).
          example: 'Pronos Ligue 1 - Journée 17'
        body_html:
          type: string
          description: >-
            The draft's current saved body markup, exactly as stored (this
            operation does not sanitise — sanitisation of writer-authored
            HTML happens at publish time; see `POST .../publish`).
          example: '<p>Notre analyse de la journée 17...</p>'
...